# manager-zedaapi/public/openapi-content-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zé da API - Content Management API",
    "description": "API completa para gerenciamento externo de conteudo da plataforma Zé da API.\n\nTodos os endpoints requerem autenticacao via header `Authorization: Bearer <CONTENT_API_KEY>`.\n\nA chave e definida na variavel de ambiente `CONTENT_API_KEY` do servidor.",
    "version": "1.0.0",
    "contact": {
      "name": "Zé da API",
      "url": "https://zedaapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://app.zedaapi.com",
      "description": "Producao"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CONTENT_API_KEY definida no .env do servidor. Minimo 32 caracteres."
      }
    },
    "schemas": {
...
      "SupportCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5scat001"
          },
          "name": {
            "type": "string",
            "example": "Primeiros Passos"
          },
          "slug": {
            "type": "string",
            "example": "primeiros-passos"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Artigos para quem esta comecando na plataforma"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "example": "book-open"
          },
          "sortOrder": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "articles": {
                "type": "integer",
                "example": 5
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "API key ausente ou invalida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "summary": "Header ausente",
                "value": {
                  "error": "Missing Authorization header"
                }
              },
              "invalid": {
                "summary": "Chave invalida",
                "value": {
                  "error": "Invalid API key"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Recurso nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not found"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito (slug duplicado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "A post with this slug already exists"
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/admin/content/support/categories": {
      "get": {
        "tags": [
          "Support Categories"
        ],
        "summary": "Listar categorias de suporte",
        "description": "Retorna todas as categorias de suporte com contagem de artigos.",
        "operationId": "listSupportCategories",
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "cm5scat001",
                      "name": "Primeiros Passos",
                      "slug": "primeiros-passos",
                      "description": "Para quem esta comecando",
                      "icon": "book-open",
                      "sortOrder": 0,
                      "_count": {
                        "articles": 5
                      }
                    },
                    {
                      "id": "cm5scat002",
                      "name": "Configuracoes",
                      "slug": "configuracoes",
                      "description": "Ajustes e personalizacoes",
                      "icon": "settings",
                      "sortOrder": 1,
                      "_count": {
                        "articles": 8
                      }
                    },
                    {
                      "id": "cm5scat003",
                      "name": "Cobranca",
                      "slug": "cobranca",
                      "description": "Planos, faturas e pagamentos",
                      "icon": "credit-card",
                      "sortOrder": 2,
                      "_count": {
                        "articles": 3
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Support Categories"
        ],
        "summary": "Criar categoria de suporte",
        "description": "Cria uma nova categoria. Icones disponiveis: help-circle, book-open, settings, zap, credit-card, shield, message-square.",
        "operationId": "createSupportCategory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Seguranca"
                  },
                  "slug": {
                    "type": "string",
                    "description": "Slug customizado (opcional, auto-gerado se omitido)",
                    "example": "seguranca"
                  },
                  "description": {
                    "type": "string",
                    "example": "Artigos sobre seguranca e privacidade"
                  },
                  "icon": {
                    "type": "string",
                    "enum": [
                      "help-circle",
                      "book-open",
                      "settings",
                      "zap",
                      "credit-card",
                      "shield",
                      "message-square"
                    ],
                    "example": "shield"
                  },
                  "sortOrder": {
                    "type": "integer",
                    "example": 4
                  }
                }
              },
              "example": {
                "name": "Seguranca",
                "description": "Artigos sobre seguranca e privacidade",
                "icon": "shield",
                "sortOrder": 4
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportCategory"
                },
                "example": {
                  "id": "cm5scat004",
                  "name": "Seguranca",
                  "slug": "seguranca",
                  "description": "Artigos sobre seguranca e privacidade",
                  "icon": "shield",
                  "sortOrder": 4,
                  "createdAt": "2026-02-21T16:00:00.000Z",
                  "updatedAt": "2026-02-21T16:00:00.000Z",
                  "_count": {
                    "articles": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nome obrigatorio",
            "content": {
              "application/json": {
                "example": {
                  "error": "Missing required field: name"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/admin/content/support/categories/{id}": {
      "patch": {
        "tags": [
          "Support Categories"
        ],
        "summary": "Atualizar categoria de suporte",
        "operationId": "updateSupportCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5scat001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "name": "Primeiros Passos - Atualizado",
                "icon": "zap",
                "sortOrder": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportCategory"
                },
                "example": {
                  "id": "cm5scat001",
                  "name": "Primeiros Passos - Atualizado",
                  "slug": "primeiros-passos-atualizado",
                  "description": "Para quem esta comecando na plataforma",
                  "icon": "zap",
                  "sortOrder": 0,
                  "createdAt": "2026-01-05T08:00:00.000Z",
                  "updatedAt": "2026-02-21T16:30:00.000Z",
                  "_count": {
                    "articles": 5
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Support Categories"
        ],
        "summary": "Excluir categoria de suporte",
        "description": "Exclui uma categoria. Falha se houver artigos vinculados.",
        "operationId": "deleteSupportCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5scat001"
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria excluida",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "400": {
            "description": "Categoria tem artigos vinculados",
            "content": {
              "application/json": {
                "example": {
                  "error": "Cannot delete category with 5 article(s). Move or delete them first."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}